requirements.txt
phidata
groq 
flask
python-dotenv
duckduckgo-search
sqlalchemy
bcrypt
auth
requests
mysql-connector-python
